Fix update and favorite
