+++ ui/src/routes/app.boards.id/contexts/useFavoriteBoardMutation.ts
@@ -1,11 +1,13 @@
 import immer from 'immer'
 import { useQueryClient } from 'react-query'
-import { useMutation, MutationReturnType } from '~/src/contexts/Query'
-import { DivdedQueryAndMutationProps } from './types'
+import { useMutation } from '~/src/contexts/Query'
+import { DivdedQueryAndMutationProps, CustomMutationReturnType } from './types'
 
 interface FavoriteBoardMutationVariables {
   is_favorite: boolean
 }
+
+type MutationReturnType = CustomMutationReturnType<Function>
 
 const useFavoriteBoardMutation = ({
   board,
@@ -14,12 +16,33 @@ const useFavoriteBoardMutation = ({
 }: DivdedQueryAndMutationProps): MutationReturnType => {
   const queryClient = useQueryClient()
 
-  return useMutation<FavoriteBoardMutationVariables>((v) => `/boards/${board?.id}/favorite`, 'put', {
-    onSuccess: (data) => {
-      // @TODO: Toast
-      queryClient.invalidateQueries('/boards')
+  const { mutate: mutateFn, ...props } = useMutation<{}, FavoriteBoardMutationVariables>(
+    (v) => `/boards/${board?.id}/favorite`,
+    'put',
+    {
+      onSuccess: () => {
+        // @TODO: Toast
+        queryClient.invalidateQueries('/boards')
+      }
     }
-  })
+  )
+
+  const mutate = () => {
+    const updatedFavoriteStatus = !board.is_favorite
+
+    setBoard(
+      immer(board, (draft) => {
+        draft.is_favorite = updatedFavoriteStatus
+      })
+    )
+
+    mutateFn({ is_favorite: updatedFavoriteStatus })
+  }
+
+  return {
+    mutate,
+    ...props
+  }
 }
 
 export { useFavoriteBoardMutation, FavoriteBoardMutationVariables }
